requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
black==23.12.0
click==8.1.7
crispy-bootstrap5==2023.10
# Django==5.0
django-crispy-forms==2.1
django-phonenumber-field==7.2.0
frozenlist==1.4.1
idna==3.6
inflect==7.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.26
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
yarl==1.9.4
